Add #7047; Update 28945
Jon Moss wrote a better reproduction script for #28945

The script for #7047 came from trying to reproduce what Katrina Owen was
using for testing. It may need to be updated. It looks like this issue
still exists, but it may be worked another way since Andrew White added
Custom Routes in 5 and he was working to make this a part of 5.